healthcare_diet_app/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# Medical Chatbot Dependencies
sentence-transformers==5.1.0
langchain==0.3.27
langchain-google-genai==2.0.10
google-generativeai==0.8.5
langchain-pinecone==0.2.11
pinecone[grpc]==7.3.0
langchain_community==0.3.29
langchain_experimental==0.3.4
pypdf==6.0.0
transformers==4.55.4
torch==2.8.0
huggingface-hub==0.34.4
